refactor: progress logging
