+++ packages/core-loggers/src/all.ts
@@ -1,4 +1,3 @@
-export * from './cliLogger'
 export * from './deprecationLogger'
 export * from './fetchingProgressLogger'
 export * from './hookLogger'
