+++ packages/core-loggers/src/index.ts
@@ -1,7 +1,6 @@
 export * from './all'
 
 import {
-  CliLog,
   DeprecationLog,
   FetchingProgressLog,
   HookLog,
@@ -20,8 +19,7 @@ import {
   SummaryLog,
 } from './all'
 
-export type Log = CliLog
-  | DeprecationLog
+export type Log = DeprecationLog
   | FetchingProgressLog
   | HookLog
   | ImportingLog
